Finished solver (sort of)
Finished solver and tested on small array for npe matter. Need to extend test to other particles. Woo hoo. Now just need to see if it works in general case with more particles... So far it takes a while to solve, would be better to leave it overnight and stuff.