Minor edits
Made some edits: note that with a closer initial guess, the broyden method solves much faster. Also it seems it is still sensitive to initial guesses. So... might be probelmatic. But we accomplished something good tonight. End on a good note. Tomorrow will try for whole system and then move on to sound speed calculations again with better data.